+++ Assembler/lexer.h
@@ -2,8 +2,9 @@
 #define _LEXER
 
 #include "token.h"
+#include <stdint.h>
 
-int initLexer(char* filename); // Lexer initialisation
+int16_t initLexer(char* filename); // Lexer initialisation
 struct Token nextToken(); // Get next token
 
 #endif
